fix type

+++ android/src/main/java/com/reactpaxlibrary/RNPaxLibraryModule.java
@@ -54,11 +54,11 @@ public class RNPaxLibraryModule extends ReactContextBaseJavaModule {
     }
 
         @ReactMethod
-    public void printText(String text, int grayLevel, short alignMode,  Double cutMode) {
+    public void printText(String text, int grayLevel, int alignMode,  Double cutMode) {
         try {
             printer.init();
             printer.setGray(grayLevel);
-            printer.setAlignMode(alignMode);
+            printer.setAlignMode((short) alignMode);
             printer.printStr(text, null);
             printer.start();
         } catch (Exception e) {
